Move MongoStoreConfig to the mongo package, where it should have been.
